makes ExecOpGenericIndexNestedLoopsJoinWithRequests an extension of ExecOpGenericIndexNestedLoopsJoinBase, and adds some documentation
